Rename commons to more fitting name: helper
